replace await and move tests
