+++ pages/Home.test.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { render, screen } from "@testing-library/react"
 
-import Home from "../../pages/index"
+import Home from "./index"
 
 describe("Home", () => {
   it("should render an H2", () => {
